[updated] update the file name and folder
