+++ src-tauri/src/apis/mod.rs
@@ -1,3 +1,4 @@
 //! This module is the collection of all widget API plugins for `@deskulpt-test/apis`.
 
 pub(crate) mod fs;
+pub(crate) mod sys;
