+++ src-tauri/src/main.rs
@@ -13,8 +13,6 @@ mod setup;
 mod states;
 mod utils;
 
-mod widget;
-
 #[cfg(test)]
 mod testing;
 
@@ -48,7 +46,7 @@ fn main() {
         .plugin(tauri_plugin_global_shortcut::Builder::new().build())
         .plugin(tauri_plugin_shell::init())
         .plugin(apis::fs::init())
-        .plugin(widget::sys::init())
+        .plugin(apis::sys::init())
         .run(tauri_build_context!())
         .expect("Error running the Deskulpt application");
 }
