+++ tooling/apis/src/sys.ts
@@ -1,6 +1,6 @@
 import { invoke } from "@tauri-apps/api/core";
 function getSystemInfo(widgetId: string, path: string) {
-  return invoke("plugin:widget-sys|get_system_info", {
+  return invoke("plugin:apis-sys|get_system_info", {
     widgetId: widgetId,
     path: path,
   });
